+++ src/FarMenu.cpp
@@ -9,7 +9,7 @@ FarMenu::FarMenu(): Flags(0), Selection(0)
 
 FarMenu::~FarMenu()
 {
-	for(int i = 0; i < items.size(); ++i) {
+	for(size_t i = 0; i < items.size(); ++i) {
 		delete(items[i].Text);
 	}
 }
